⚙️ EditSessionForm refactor
